Move prod types to the EssentialFeed production module
